Post-catch up meeting code changes - table of contents correctly extracted out to component; TOC template markup refactored; hero/footer content-bottom adjustments
