+++ src/styles/main.scss
@@ -150,7 +150,7 @@
     margin: 3rem 0 0 0;
 }
 
-.hero.footer .content-bottom {
+.hero .content-bottom {
     justify-content: space-between;
 
     p:first-of-type {
